Very minor fixes  * Make sure to correctly cast abort parameter to a string  * Move log to `Error` namespace  * Remove debug log from `ScriptService`
