+++ lib/Interpreter/FunctionProvider.php
@@ -3,6 +3,7 @@
 namespace OCA\FilesScripts\Interpreter;
 
 use OCA\FilesScripts\Interpreter\Functions\Error\Abort;
+use OCA\FilesScripts\Interpreter\Functions\Error\Log;
 use OCA\FilesScripts\Interpreter\Functions\Files\Copy_File;
 use OCA\FilesScripts\Interpreter\Functions\Files\Directory_Listing;
 use OCA\FilesScripts\Interpreter\Functions\Files\Exists;
@@ -43,7 +44,6 @@ use OCA\FilesScripts\Interpreter\Functions\Util\Http_Request;
 use OCA\FilesScripts\Interpreter\Functions\Util\Json;
 use OCA\FilesScripts\Interpreter\Functions\Util\Sort;
 use OCA\FilesScripts\Interpreter\Functions\Util\Wait;
-use OCA\FilesScripts\Interpreter\Functions\Util\Log;
 
 class FunctionProvider implements IFunctionProvider {
 	/** @var RegistrableFunction[] */
